🎨 - Refactored namespaces
Moved a few of the APIs to be nested under things that make sense.
