+++ lib/db/dbopts.js
