+++ lib/db/icv.js
@@ -1,7 +1,7 @@
 const { fetch } = require('fetch-ponyfill')();
 const qs = require('querystring');
 
-const { httpBody, httpMessage } = require('./response-transforms');
+const { httpBody, httpMessage } = require('../response-transforms');
 
 const get = (conn, database, options, params = {}) => {
   const headers = conn.headers();
