+++ lib/db/index.js
@@ -1,0 +1,6 @@
+const db = require('./main');
+const icv = require('./icv');
+const transaction = require('./transaction');
+const options = require('./options');
+
+module.exports = Object.assign({}, db, { icv }, { transaction }, { options });
