+++ lib/db/main.js
@@ -4,17 +4,8 @@ const qs = require('querystring');
 const flat = require('flat');
 const lodashGet = require('lodash/get');
 
-const DB_OPTIONS = require('./dbopts');
-const { httpMessage, httpBody } = require('./response-transforms');
-
-const dispatchDBOptions = (conn, config, body) => {
-  config.headers.append('Content-Type', 'application/json');
-  return fetch(conn.uri('admin', 'databases', config.database, 'options'), {
-    method: config.method,
-    headers: config.headers,
-    body: JSON.stringify(flat(body)),
-  });
-};
+const { httpMessage, httpBody } = require('../response-transforms');
+const { get: getOptions } = require('./options');
 
 const create = (conn, database, databaseOptions = {}, options = {}, params) => {
   const headers = conn.headers();
@@ -51,41 +42,6 @@ const getDatabase = (conn, database, params) => {
   return fetch(conn.uri(database), {
     headers,
   }).then(httpBody);
-};
-
-const getOptions = (conn, database, params) => {
-  const headers = conn.headers();
-  return dispatchDBOptions(
-    conn,
-    {
-      headers,
-      database,
-      method: 'PUT',
-    },
-    DB_OPTIONS
-  )
-    .then(httpBody)
-    .then(res => {
-      if (res.status === 200) {
-        return Object.assign({}, res, {
-          body: flat.unflatten(res.body),
-        });
-      }
-      return res;
-    });
-};
-
-const setOptions = (conn, database, databaseOptions, params) => {
-  const headers = conn.headers();
-  return dispatchDBOptions(
-    conn,
-    {
-      headers,
-      database,
-      method: 'POST',
-    },
-    databaseOptions
-  ).then(httpMessage);
 };
 
 const offline = (conn, database, params) => {
@@ -199,8 +155,6 @@ module.exports = {
   create,
   drop,
   get: getDatabase,
-  getOptions,
-  setOptions,
   offline,
   online,
   optimize,
