+++ lib/db/transaction.js
@@ -1,8 +1,8 @@
 const { fetch } = require('fetch-ponyfill')();
 const qs = require('querystring');
 
-const { httpBody, httpMessage } = require('./response-transforms');
-const { mimeType, buildQuery } = require('./query/utils');
+const { httpBody, httpMessage } = require('../response-transforms');
+const { mimeType, buildQuery } = require('../query/utils');
 
 const attachTransactionId = transactionId => res =>
   Object.assign({}, res, { transactionId });
