+++ lib/index.js
@@ -4,11 +4,8 @@ const Stardog = {
   Connection: require('./Connection'),
   db: require('./db'),
   query: require('./query'),
-  server: require('./server'),
-  icv: require('./icv'),
   user: require('./user'),
-  role: require('./role'),
-  transaction: require('./transaction'),
+  server: require('./server'),
 };
 
 module.exports = Stardog;
