+++ lib/user/index.js
@@ -1,0 +1,4 @@
+const user = require('./main');
+const role = require('./role');
+
+module.exports = Object.assign({}, user, { role });
