+++ lib/user/main.js
@@ -1,5 +1,5 @@
 const { fetch } = require('fetch-ponyfill')();
-const { httpBody, httpMessage } = require('./response-transforms');
+const { httpBody, httpMessage } = require('../response-transforms');
 
 const list = (conn, params) => {
   const headers = conn.headers();
